requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.3
pyaml==25.7.0
PyYAML==6.0.3
